Update UnitsSystem.lean
In the docstring for Scalar.neZero_iff, a comment mentions the possibility of an API for positive scalars. This PR implements that idea, with a focus on enabling functions like log and exp, as awell as a single, general Scalar.map (f : ℝ → ℝ) function that lifts any real function to dimensionless scalars.